// generated/type-graphql/resolvers/inputs/PatientsWhereUniqueInput.ts
export class PatientsWhereUniqueInput {
  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  id?: string | undefined;

  @TypeGraphQL.Field(_type => TypeGraphQL.Int, {
    nullable: true
  })
  cin?: number | undefined;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  general_medical_record_id?: string | undefined;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  user_id?: string | undefined;

  @TypeGraphQL.Field(_type => [PatientsWhereInput], {
    nullable: true
  })
  AND?: PatientsWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [PatientsWhereInput], {
    nullable: true
  })
  OR?: PatientsWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [PatientsWhereInput], {
    nullable: true
  })
  NOT?: PatientsWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  date_of_birth?: DateTimeFilter | undefined;

  @TypeGraphQL.Field(_type => Enumpatients_gender_enumFilter, {
    nullable: true
  })
  gender?: Enumpatients_gender_enumFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  profile_image?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => GeneralMedicalRecordsNullableRelationFilter, {
    nullable: true
  })
  GeneralMedicalRecords_GeneralMedicalRecords_patient_idTopatients?: GeneralMedicalRecordsNullableRelationFilter | undefined;

  @TypeGraphQL.Field(_type => AuthorizationsListRelationFilter, {
    nullable: true
  })
  authorizations?: AuthorizationsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => CertificatesListRelationFilter, {
    nullable: true
  })
  certificates?: CertificatesListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => ConsultationsListRelationFilter, {
    nullable: true
  })
  consultations?: ConsultationsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Lab_documentsListRelationFilter, {
    nullable: true
  })
  lab_documents?: Lab_documentsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Lab_requestsListRelationFilter, {
    nullable: true
  })
  lab_requests?: Lab_requestsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => UsersNullableRelationFilter, {
    nullable: true
  })
  users?: UsersNullableRelationFilter | undefined;

  @TypeGraphQL.Field(_type => GeneralMedicalRecordsNullableRelationFilter, {
    nullable: true
  })
  GeneralMedicalRecords_patients_general_medical_record_idToGeneralMedicalRecords?: GeneralMedicalRecordsNullableRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PrescriptionsListRelationFilter, {
    nullable: true
  })
  prescriptions?: PrescriptionsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Rdv_requestsListRelationFilter, {
    nullable: true
  })
  rdv_requests?: Rdv_requestsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => RdvsListRelationFilter, {
    nullable: true
  })
  rdvs?: RdvsListRelationFilter | undefined;
}